backend/pyproject.toml
[tool.poetry]
name = "frzn-docs-backend"
version = "0.1.0"
description = "FastAPI + SQLModel backend for frzn-docs"
authors = ["Arnav Motwani <arnavmotwani77@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.11"
fastapi = ">=0.115.12,<0.116.0"
uvicorn = { extras = ["standard"], version = "^0.23.0" }
sqlmodel = ">=0.0.24,<0.0.25"
psycopg2-binary = ">=2.9.10,<3.0.0"
alembic = ">=1.16.1,<2.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
pgvector = ">=0.4.1,<0.5.0"
pydantic-settings = "^2.2.1"
gitpython = "^3.1.44"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"